My commit to build a guess game.
This program allows you to make "A Guess" by asking for your input. It allows you to type in maximum of 3 guesses, if after the third guess, you do not get the secret word, it prints to your screen "Out of Guesses, YOU LOSE!". If you got the secret word spelt correctly before running of out_of_guesses, then it prints "Wow! You just won an outfit for guessing right."